requirements.txt
pandas
numpy
matplotlib
pyyaml
pyproj
folium
requests
shapely
geopandas
pyogrio
